added security

+++ main.py
@@ -67,12 +67,16 @@ def update_task(task_id):
 
 
 # DELETE task
-@app.route('/tasks/<int:task_id>', methods=['POST'])
+@app.route('/tasks/<int:task_id>', methods=['POST','DELETE'])
 def delete_task(task_id):
-    task = Task.query.get_or_404(task_id)
-    db.session.delete(task)
-    db.session.commit()
-    return redirect(url_for('home'))
+    appkey = request.args.get("appkey")
+    if appkey == "websitesecret":
+        task = Task.query.get_or_404(task_id)
+        db.session.delete(task)
+        db.session.commit()
+        return redirect(url_for('home'))
+    else:
+        return render_template("error.html")
 
 
 if __name__ == '__main__':
